Restructure code
Previously, too much logical code was duplicate and stuffed into the IncomingServer.php. This made it difficult to handle the recursive nature of multiparts that may or may not have multiparts in themselves. It also made header VS body (in the DATA) hard to parse. This is an improved version